+++ v2/src/dar_backup/manager.py
@@ -535,7 +535,6 @@ def main():
 
     args.config_file = os.path.expanduser(os.path.expandvars(args.config_file))
     config_settings = ConfigSettings(args.config_file)
-    print(f"Config settings: {config_settings}")
 
     if not os.path.dirname(config_settings.logfile_location):
         print(f"Directory for log file '{config_settings.logfile_location}' does not exist, exiting")